requirements.txt
numpy==1.24.2
Pillow==8.1.1
scikit_learn==1.2.2
scipy==1.10.1
torch==2.1.0
torchvision==0.16.0
tqdm==4.32.1

